pyproject.toml
[project]
name = "pyani-benchmark"
version = "0.0.1"
authors = [
    { name = "Leighton Pritchard", email = "leighton.pritchard@strath.ac.uk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "biopython",
    "networkx",
    "numpy",
    "rich",
    "typer",
]

[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pyani_benchmark"]

[project.scripts]
"pyani-benchmark" = "pyani_benchmark.script:app"

[project.urls]
Homepage = "https://github.com/pyani-plys/benchmark"
Issues = "https://github.com/pyani-plys/benchmark/issues"